pyproject.toml
[tool.poetry]
name = "act"
version = "0.0.0.post2.dev0+f632f4e"
description = "act - A simple task runner."
authors = [
    "נυαη נυαηѕση <@retr0h>",
]
maintainers = [
    "נυαη נυαηѕση <@retr0h>",
]
readme = "README.md"
homepage = "https://github.com/retr0h/act"
repository = "https://github.com/retr0h/act"
documentation = "https://github.com/retr0h/act"
packages = [
    { include = "act" },
]

[tool.poetry.dependencies]
python = "^3.6"
click = "*"
click-completion = "*"
click-help-colors = "*"
toposort = "*"
PyYAML = "*"
subprocess_tee = "*"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = {extras = ["toml"], version = "*"}
enrich = "^1.2.6"
flake8 = "*"
#flake8-annotations = "*"
hacking = "*"
isort = "*"
magicattr = "*"
poethepoet = "*"
pyfakefs = "*"
pytest = "^6.0.1"
pytest-clarity = "*"
pytest-cov = "*"
pytest-mock = "*"

[tool.poetry.scripts]
act = "act.shell:main"

[tool.black]
line-length = 79
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = [
    "test/",
]
addopts = "-v -rxXs"

[tool.coverage.run]
omit = [
    "act/cmd/*",
    "act/shell.py",
]
relative_files = true

[tool.coverage.report]
fail_under = 100

[tool.isort]
profile = "open_stack"

[tool.poetry-dynamic-versioning]
enable = false
style = "pep440"

[tool.poetry-dynamic-versioning.substitution]
files = ["act/__init__.py"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
lint = "flake8"
unit = "py.test -vv --junitxml=xunit-reports/xunit-result-act.xml"
cov = "py.test --cov act --cov-report term-missing --cov-report xml:coverage-reports/coverage-act.xml --showlocals"
format = [
    {cmd = "black ."},
    {cmd = "isort ."},
]
style = [
    {cmd = "black . --check --diff"},
    {cmd = "isort . --check-only --diff"},
]
test = ["lint", "style", "cov"]
